update session and increase/decrease quantity when delete order detail

+++ ShoppingAssignment_SE151295/ShoppingAssignment_SE151295/Pages/OrderDetails/Delete.cshtml.cs
@@ -2,6 +2,7 @@ using System;
 using System.Collections.Generic;
 using System.Linq;
 using System.Threading.Tasks;
+using Microsoft.AspNetCore.Http;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.AspNetCore.Mvc.RazorPages;
 using Microsoft.EntityFrameworkCore;
@@ -21,6 +22,9 @@ namespace ShoppingAssignment_SE151295.Pages.OrderDetails
         [BindProperty]
         public OrderDetail OrderDetail { get; set; }
 
+        [BindProperty]
+        public int oldQuanity {set;get;}
+
         public async Task<IActionResult> OnGetAsync(string orderid, int productid)
         {
             if (orderid == null)
@@ -38,28 +42,38 @@ namespace ShoppingAssignment_SE151295.Pages.OrderDetails
             {
                 return NotFound();
             }
+
             return Page();
         }
 
         public async Task<IActionResult> OnPostAsync()
         {
-           // if (id == null)
-           // {
-           //     return NotFound();
-           // }
-
-            //OrderDetail = await _context.OrderDetails.FindAsync(id);
+            if(string.IsNullOrEmpty(HttpContext.Session.GetString("username")))
+            {
+                return RedirectToPage("/Login/MyLogin","Session");
+            }
 
             string orderid = "";
 
             if (OrderDetail != null)
             {
                 orderid = OrderDetail.OrderId;
+
+                OrderDetail tmepOrderDetail =  _context.OrderDetails
+                            .Include(o => o.Order)
+                            .Include(o => o.Product)
+                            .Where(p => p.OrderId == OrderDetail.OrderId && p.ProductId == OrderDetail.ProductId)
+                            .SingleOrDefault();
+
+                int shiftQuantity = tmepOrderDetail.Quantity;
+                Product product = _context.Products.Where(p => p.ProductId == OrderDetail.ProductId).SingleOrDefault();
+                product.QuantityPerUnit = product.QuantityPerUnit + shiftQuantity;
+                _context.Update(product);
                 _context.OrderDetails.Remove(OrderDetail);
                 await _context.SaveChangesAsync();
             }
 
-            return RedirectToPage("./OrderDetailManage",new {id = orderid});
+            return RedirectToPage("./OrderDetailManage",new {id = orderid.Trim()});
         }
     }
 }
